requirements.txt
google-adk[eval]
google-cloud-aiplatform[agent_engine]
fastapi
uvicorn[standard]
python-dotenv
firebase-admin

